test: TETROMINO RECT の E2E テストを game.spec.js に追加 (#134)
既存テストにTETROMINO RECTのカバレッジがなかったため、
以下4テストをPUZZLEモードセクションに追加:
- イントロ表示→ゲーム開始
- ピース選択・解除
- カラー切り替え
- HOMEでタイトル復帰
